content-generator-function/requirements.txt
functions-framework>=3.0.0
google-cloud-bigquery>=3.10.0
google-cloud-storage>=2.10.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
